Weekly Logs Fully Implemented
+ Can now log to one file for one week. This means that a file will be
created every monday morning for the new week. If the program
crashes/gets closed we can restart the program and select "Start Day" in
admin window. This will allow us to select an existing csv file to write
to.
+ Removed number_records stat since unneeded
+ If a customer selects "Rental" then it will take them to the summary
page, skipping the problem page.
+ fixed bug in description text box where a user could press backspace
or space and it would not auto delete default text
+ Changed start/end time to 7:30/5:30 respectively
+ Changed time format in csv file to show date and military time

+++ CustomerLogger/CSVWriter.cs
@@ -24,16 +24,22 @@ namespace CSV {
         private StreamWriter _sw;
         private string _fname;
 
-        public CSVWriter(string fname) {
+        /// <summary>
+        /// Creates a new csv writer. 
+        /// Can create a new file or append to existing
+        /// </summary>
+        /// <param name="fname">CSV File to write to</param>
+        /// <param name="mode">How to initially handle file, Create will delete an existing file. Append will not</param>
+        public CSVWriter(string fname, FileMode mode) {
 
             _line = "";
             _fname = fname;
 
-            if(true == File.Exists(_fname)) { 
-            
-                File.Delete(_fname);
+            if (mode == FileMode.Create) {
+                if (true == File.Exists(_fname)) {
+                    File.Delete(_fname);
+                }
             }
-
         }
 
         public void addToCurrent(string text) {
@@ -56,6 +62,11 @@ namespace CSV {
             }
         }
 
+        /// <summary>
+        /// Opens _fname
+        /// Appends a line to the file
+        /// Closes file
+        /// </summary>
         public void WriteLine() {
 
             using(_fs = new FileStream(_fname, FileMode.Append))
